refactors apps
