+++ data_copilot/celery_app/apps/prompt_execution_app.py
@@ -5,15 +5,12 @@ from typing import Tuple
 import openai
 from celery import Celery, chain
 from celery.exceptions import SoftTimeLimitExceeded, TimeLimitExceeded
-from langchain.agents import create_pandas_dataframe_agent
-from langchain.llms import OpenAI
-from openai import error
 
 from data_copilot.celery_app.config import Config
 from data_copilot.celery_app.crud.chats import crud_create_message
 from data_copilot.celery_app.database.psql import SessionLocal, engine
-from data_copilot.celery_app.executors import helpers
 from data_copilot.db_models.base import Base
+from data_copilot.execution_apps import get_app
 
 Base.metadata.create_all(bind=engine)
 
@@ -23,7 +20,6 @@ CONFIG = Config()
 execution_app = Celery("main", broker=CONFIG.CELERY_BROKER_URL)
 
 
-# retry 2 times in case of failure
 @execution_app.task(
     name="save_result",
     soft_time_limit=10,
@@ -34,20 +30,17 @@ def save_result(
     prompt_result: Tuple[str, str],
     chat_id: uuid.UUID,
     artifact_version_id: uuid.UUID | None = None,
-    sas_url: str | None = None,
-    artifact_config: dict | None = None,
     message_id: uuid.UUID | None = None,
 ):
     """Save the final result as message in the DB.
 
     Args:
-        prompt_result (Tuple[str, str]): _description_
-        chat_id (uuid.UUID): _description_
-        artifact_version_id (uuid.UUID | None, optional): _description_.
-        Defaults to None.
-        sas_url (str | None, optional): _description_. Defaults to None.
-        artifact_config (dict | None, optional): _description_. Defaults to None.
-        message_id (uuid.UUID | None, optional): _description_. Defaults to None.
+        prompt_result (Tuple[str, str]): Tuple of message type and message content.
+        chat_id (uuid.UUID): Chat id of the message.
+        artifact_version_id (uuid.UUID, optional): Artifact version id of the message.
+            Defaults to None.
+        message_id (uuid.UUID, optional): Message id of the message. Defaults to None.
+
     """
     try:
         message_type, message_content = prompt_result
@@ -80,40 +73,38 @@ def save_result(
     autoretry_for=(TimeLimitExceeded, openai.error.RateLimitError),
     retry_kwargs={"max_retries": 3, "countdown": 30},
 )
-def translate_user_prompt(
+def execute_user_prompt(
     user_prompt: str,
+    chat_id: uuid.UUID,
     message_id: uuid.UUID,
-    artifact_config: dict | None = None,
-    sas_url: str | None = None,
+    artifact_version_id: uuid.UUID,
+    artifact_version_uri: str,
 ) -> Tuple[str, str]:
     """Translate user prompt into method to be called
 
     Args:
-        user_prompt (str): Prompt passed from the user message to be executed.
-        message_id (uuid.UUID): message id of the prompt.
-        artifact_config (dict):
+        user_prompt (str): User prompt to be translated.
+        chat_id (uuid.UUID): Chat id of the message.
+        message_id (uuid.UUID): Message id of the message.
+        artifact_version_id (uuid.UUID): Artifact version id of the message.
+        artifact_version_uri (str): Artifact version uri of the message.
 
     Returns:
         Tuple[str, str]: Tuple of message type and message content.
     """
+
     try:
-        # schema = artifact_config.get("files", [dict()])[0].get("file_schema", {})
-        file_type = artifact_config.get("files", [dict()])[0].get("file_type", "")
-
-        dataset = helpers.read_dataset(sas_url, file_type)
-
-        if len(dataset.index) == 0:
-            raise Exception(f"Wrong '{sas_url}' content")
-
-        agent = create_pandas_dataframe_agent(
-            OpenAI(temperature=0), dataset, verbose=False
+        result = (
+            get_app()
+            .execute_message(
+                user_prompt=user_prompt,
+                chat_id=chat_id,
+                message_id=message_id,
+                artifact_version_id=artifact_version_id,
+                artifact_version_uri=artifact_version_uri,
+            )
+            .to_dict()
         )
-
-        answer = agent.run(user_prompt)
-        message = helpers.Message(helpers.MessageTypes.TEXT, "Answer")
-        message.add_text(f"{answer}")
-        result = message.to_dict()
-        print(result)
         return result["message_type"], result["text_content"]
 
     except SoftTimeLimitExceeded:
@@ -123,35 +114,14 @@ def translate_user_prompt(
             f"message_id: {message_id}"
         )
         raise TimeLimitExceeded()
-
-    except error.RateLimitError:
-        logging.error(
-            "The translation of the user prompt failed due to rate limit error --"
-            f"Prompt: {user_prompt} --"
-            f"message_id: {message_id}"
-        )
-        return "error", "Rate limit error from OpenAI API. Please try again later"
-
-    except error.AuthenticationError:
-        logging.error(
-            "The translation of the user prompt failed due to authentication error --"
-            f"Prompt: {user_prompt} --"
-            f"message_id: {message_id}"
-        )
-        return "error", "OpenAI API key is invalid"
-
     except Exception as e:
         logging.error(
             f"An error occured while translating the user prompt: {e} --"
             f"Prompt: {user_prompt} --"
             f"message_id: {message_id}"
         )
-
         logging.exception(e)
-        return (
-            "error",
-            "An error occured while executing. Please have a look at the logs.",
-        )
+        return "text", "An error occured while translating the user prompt."
 
 
 @execution_app.task(name="execute_user_message")
@@ -160,21 +130,19 @@ def execute_user_message(
     chat_id: uuid.UUID,
     message_id: uuid.UUID,
     artifact_version_id: uuid.UUID | None = None,
-    sas_url: str | None = None,
-    artifact_config: dict | None = None,
+    artifact_version_uri: str | None = None,
 ):
     chain(
-        translate_user_prompt.s(
-            user_prompt,
+        execute_user_prompt.s(
+            user_prompt=user_prompt,
+            chat_id=chat_id,
             message_id=message_id,
-            artifact_config=artifact_config,
-            sas_url=sas_url,
+            artifact_version_id=artifact_version_id,
+            artifact_version_uri=artifact_version_uri,
         ),
         save_result.s(
-            chat_id,
-            artifact_version_id,
-            sas_url,
-            artifact_config,
+            chat_id=chat_id,
+            artifact_version_id=artifact_version_id,
             message_id=message_id,
         ),
     )()
