+++ data_copilot/celery_app/config/config.py
@@ -6,7 +6,6 @@ class Config(BaseSettings):
 
     # Server
     CELERY_BROKER_URL: str = Field(..., env="CELERY_BROKER_URL")
-    OPENAI_API_KEY: str = Field(..., env="OPENAI_API_KEY")
     DB_CONNECTION_STRING: str = Field(..., env="DB_CONNECTION_STRING")
 
     @property
