+++ pyproject.toml
@@ -53,6 +53,16 @@ dependencies = [
   "xlrd>=2.0.1"
 ]
 
+[project.optional-dependencies]
+dev = [
+    'black >= 23.3.0, <= 23.3.0',
+    'click>=8.1.3',
+    'ddt>=1.6.0',
+    'pytest>=7.2.2',
+    'tox>=4.4.8',
+    'xlwt>=1.3.0'
+]
+
 [project.scripts]
   data-copilot = "data_copilot.main:main"
