--- requirements_dev.txt
@@ -1,6 +1,0 @@
-black>=23.3.0
-click>=8.1.3
-ddt>=1.6.0
-pytest>=7.2.2
-tox>=4.4.8
-xlwt>=1.3.0
